+++ src/tensors/cpu/backend.h
@@ -10,17 +10,11 @@ namespace marian {
 namespace cpu {
 
 class Backend : public marian::Backend {
-protected:
-  bool optimized_{false};
-
 public:
   Backend(DeviceId deviceId, size_t seed) : marian::Backend(deviceId, seed) {}
   void setDevice() override {}
   void synchronize() override {}
-
-  // for CPU & inference only, sets to use optimized code for inference. Does nothing for GPU.
-  void setOptimized(bool optimize) override { optimized_ = optimize; }
-  bool isOptimized() override { return optimized_; }
 };
+
 }  // namespace cpu
 }  // namespace marian