+++ src/tensors/cpu/fbgemm/expanded_gemm.h
@@ -2,7 +2,7 @@
 
 #include "graph/node.h"
 #include "packed_gemm.h"
-#include "tensors/cpu/sharp/int_gemm.h"
+#include "tensors/cpu/integer_common.h"
 
 #if USE_FBGEMM
 #ifdef __GNUC__
@@ -57,14 +57,12 @@ struct FbgemmPacked16PackNodeOp : public UnaryNodeOp {
   int nbcol_;
   uint64_t packsize_;
 
-  FbgemmPacked16PackNodeOp(Expr a, PackMatrix packMat, bool transpose, float clipValue)
+  FbgemmPacked16PackNodeOp(Expr a, PackMatrix packMat, bool transpose)
       : UnaryNodeOp(a, newShape(a, transpose), Type::uint8),
         packMat_(packMat),
         transpose_(transpose) {
     if(packMat != PackMatrix::B)
       ABORT("Only prepacking of B (weight matrix) is supported");
-    if(clipValue != 0)
-      ABORT("Clipping is not supported");
     if(!memoize_)
       ABORT("Only constant weight node can be packed");
   }
@@ -144,16 +142,13 @@ struct FbgemmPacked8PackNodeOp : public UnaryNodeOp {
   FbgemmPacked8PackNodeOp(Expr a,
                           PackMatrix packMat,
                           marian::Type packType,
-                          bool transpose,
-                          float clipValue)
+                          bool transpose)
       : UnaryNodeOp(a, newShape(a, transpose), Type::uint8),
         packMat_(packMat),
         packType_(packType),
         transpose_(transpose) {
     if(packMat != PackMatrix::B)
       ABORT("Only prepacking of B (weight matrix) is supported");
-    if(clipValue != 0)
-      ABORT("Clipping is not supported");
     if(!memoize_)
       ABORT("Only constant weight node can be packed");
   }
@@ -337,7 +332,7 @@ public:
                                            k_,
                                            transA_,
                                            transB_);
-                       marian::cpu::int16::AddBias(val_, child(2)->val())) };
+                       marian::cpu::integer::AddBias(val_, child(2)->val())) };
     } else {
       nodeOps = { NodeOp(fbgemmPacked8Gemm(val_,
                                            child(0)->val(),
@@ -377,11 +372,11 @@ static inline Expr affine(Expr a, Expr b, Shape bShape, Expr c, bool transA, boo
   }
 }
 
-static inline Expr pack(Type elementType, Expr a, PackMatrix packMat, bool transpose, float clipValue) {
+static inline Expr pack(Type elementType, Expr a, PackMatrix packMat, bool transpose) {
   if (elementType == Type::packed16)
-    return Expression<FbgemmPacked16PackNodeOp>(a, packMat, transpose, clipValue);
+    return Expression<FbgemmPacked16PackNodeOp>(a, packMat, transpose);
   else if (isPacked(elementType) && sizeOf(elementType) == 1)
-    return Expression<FbgemmPacked8PackNodeOp>(a, packMat, elementType, transpose, clipValue);
+    return Expression<FbgemmPacked8PackNodeOp>(a, packMat, elementType, transpose);
   else {
     ABORT("Only int8 and fp16 are available. {}", elementType);
     return nullptr;