+++ src/training/graph_group_singleton.h
@@ -35,7 +35,6 @@ public:
     graph_ = New<ExpressionGraph>();
     graph_->setDevice(deviceId);
     graph_->setCheckpointing(options_->get<bool>("gradient-checkpointing"));
-    graph_->getBackend()->setClip(options_->get<float>("clip-gemm"));
     graph_->reserveWorkspaceMB(options_->get<size_t>("workspace"));
     opt_ = Optimizer(options_);
     builder_ = models::createCriterionFunctionFromOptions(options_, models::usage::training);